+++ lib/data/send_reply.dart
@@ -1,8 +1,7 @@
 import 'package:http/http.dart' as http;
-import 'package:investops/page/login.dart';
 
 Future<String> sendReply(String reply, String id) async {
-  var url = Uri.parse('https://investops.up.railway.app/suggestionbox/replyFeedback/'+id+'/');
+  var url = Uri.parse('https://investops.up.railway.app/suggestionbox/replyFeedback/$id/');
   // var url = Uri.parse('http://localhost:8000/suggestionbox/replyFeedback/'+id+'/');
   Map<String, String> requestBody = {
     'reply': reply,
