+++ lib/model/suggestion_box.dart
