+++ lib/page/crypto/crypto_form.dart
@@ -173,13 +173,14 @@ class _MyCryptoFormPageState extends State<MyCryptoFormPage> {
                 ),
                 onPressed: () async {
                   if (_formKey.currentState!.validate()) {
-                    final response =
-                        await request.post("${siteUrl}/crypto/add/", {
+                    
+                        await request.post("$siteUrl/crypto/add/", {
                       'kode_crypto': kodeCrypto,
                       'nama_crypto': namaCrypto,
                       'harga_crypto': '$hargaCrypto',
                       'risk': risk,
-                    });
+                    }).then((value) {
+
                     _formKey.currentState!.reset();
 
                     Navigator.pushReplacement(
@@ -187,6 +188,7 @@ class _MyCryptoFormPageState extends State<MyCryptoFormPage> {
                       MaterialPageRoute(
                           builder: (context) => const MyCryptoWatchList()),
                     );
+                    });
                   }
                 },
                 child: const Text(
