+++ lib/page/drawer.dart
@@ -1,11 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:investops/assets/constants.dart';
-import 'package:investops/main.dart';
 import 'package:investops/page/login.dart';
 import 'package:investops/page/mainpage.dart';
 import 'package:investops/page/stock/mywatchlist.dart';
 import 'package:investops/page/crypto/crypto_watchlist.dart';
-import 'package:investops/page/suggestionBoxPage.dart';
+import 'package:investops/page/suggestion_box_page.dart';
 import 'package:pbp_django_auth/pbp_django_auth.dart';
 import 'package:provider/provider.dart';
 
@@ -50,11 +49,11 @@ class _UniversalDrawerState extends State<UniversalDrawer> {
               minLeadingWidth: 10,
               leading: const Icon(
                 Icons.login_rounded,
-                color: const Color.fromARGB(255, 150, 252, 3),
+                color: Color.fromARGB(255, 150, 252, 3),
               ),
               title: const Text(
                 'Login',
-                style: TextStyle(color: const Color.fromARGB(255, 150, 252, 3)),
+                style: TextStyle(color: Color.fromARGB(255, 150, 252, 3)),
               ),
               onTap: () {
                 Navigator.pushReplacement(
@@ -183,13 +182,15 @@ class _UniversalDrawerState extends State<UniversalDrawer> {
               ),
               onTap: () async {
                 if (request.loggedIn) {
-                  final response =
-                      await request.logout("$siteUrl/authenticate/logout/");
+                  
+                      await request.logout("$siteUrl/authenticate/logout/").then((value){
                   nama = '';
                   Navigator.pushReplacement(
                     context,
                     MaterialPageRoute(builder: (context) => const LoginPage()),
                   );
+
+                      });
                 }
               },
             ),
