+++ lib/page/mainpage.dart
@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:investops/page/drawer.dart';
 import 'package:provider/provider.dart';
 import 'package:pbp_django_auth/pbp_django_auth.dart';
-import 'package:flutter_staggered_grid_view/flutter_staggered_grid_view.dart';
 import 'package:countup/countup.dart';
 
 class MyMainPage extends StatefulWidget {
@@ -63,7 +62,7 @@ class _MyMainPageState extends State<MyMainPage> {
               // brightness: Brightness.light,
               // primaryColor: Color(0xff5D4524),
               primarySwatch: Colors.lime,
-              iconTheme: IconThemeData(color: Colors.white)),
+              iconTheme: const IconThemeData(color: Colors.white)),
           home: const MyHomePage(),
         ));
   }
@@ -79,21 +78,8 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  int _counter = 0;
 
-  void _incrementCounter() {
-    setState(() {
-      _counter++;
-    });
-  }
 
-  void _decrementCounter() {
-    setState(() {
-      if (_counter != 0) {
-        _counter--;
-      }
-    });
-  }
 
   @override
   Widget build(BuildContext context) {
@@ -102,10 +88,10 @@ class _MyHomePageState extends State<MyHomePage> {
           backgroundColor: Colors.black,
           title: Text(
             widget.title,
-            style: TextStyle(fontWeight: FontWeight.bold),
+            style: const TextStyle(fontWeight: FontWeight.bold),
           ),
         ),
-        drawer: UniversalDrawer(),
+        drawer: const UniversalDrawer(),
         body: Center(
         child: DraggableScrollableSheet(
             initialChildSize: 0.85,
