+++ lib/page/stock/porto_detail.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: unnecessary_brace_in_string_interps
+
 import 'package:investops/assets/constants.dart';
 import 'package:investops/page/stock/myporto.dart';
 import 'package:flutter/material.dart';
@@ -5,11 +7,11 @@ import 'package:pbp_django_auth/pbp_django_auth.dart';
 import 'package:provider/provider.dart';
 
 class PortoDetail extends StatelessWidget {
-  final kodeSaham;
-  final namaPerusahaan;
-  final hargaSaham;
-  final risk;
-  final pk;
+  final String kodeSaham;
+  final String namaPerusahaan;
+  final int hargaSaham;
+  final String risk;
+  final int pk;
 
   const PortoDetail({
     Key? key,
@@ -53,7 +55,7 @@ class PortoDetail extends StatelessWidget {
                         leading: CircleAvatar(
                           backgroundColor: Colors.white,
                           backgroundImage: NetworkImage(
-                              "https://www.idx.co.id/Portals/0/StaticData/ListedCompanies/LogoEmiten/${kodeSaham}.jpg"),
+                              "https://www.idx.co.id/Portals/0/StaticData/ListedCompanies/LogoEmiten/$kodeSaham.jpg"),
                         ),
                         title: Text(
                           kodeSaham,
@@ -101,7 +103,7 @@ class PortoDetail extends StatelessWidget {
                                             ),
                                           ),
                                           Text(
-                                            '${hargaSaham}',
+                                            '$hargaSaham',
                                             style: const TextStyle(
                                                 color: Colors.white,
                                                 fontSize: 14),
@@ -153,13 +155,13 @@ class PortoDetail extends StatelessWidget {
                                             ),
                                           ),
                                           (risk == 'LOW')
-                                              ? Text('${risk}',
+                                              ? Text(risk,
                                                   style: const TextStyle(
                                                       color: Color.fromARGB(
                                                           255, 150, 252, 3),
                                                       fontSize: 14))
                                               : Text(
-                                                  '${risk}',
+                                                  risk,
                                                   style: const TextStyle(
                                                       color: Colors.red,
                                                       fontSize: 14),
@@ -214,7 +216,7 @@ class PortoDetail extends StatelessWidget {
                     foregroundColor: Colors.black,
                   ),
                   onPressed: () async {
-                    final response = await request
+                    await request
                         .post("${siteUrl}/stock/delete_stock/${pk}", {});
                     showDialog(
                         context: context,
@@ -226,38 +228,36 @@ class PortoDetail extends StatelessWidget {
                               borderRadius: BorderRadius.circular(10),
                             ),
                             elevation: 10,
-                            child: Container(
-                              child: ListView(
-                                  padding: const EdgeInsets.only(
-                                      top: 20, bottom: 20),
-                                  shrinkWrap: true,
-                                  children: <Widget>[
-                                    const Center(
-                                        child: Text(
-                                      'Penjualan Berhasil',
+                            child: ListView(
+                                padding: const EdgeInsets.only(
+                                    top: 20, bottom: 20),
+                                shrinkWrap: true,
+                                children: <Widget>[
+                                  const Center(
+                                      child: Text(
+                                    'Penjualan Berhasil',
+                                    style: TextStyle(
+                                        fontSize: 15, color: Colors.white),
+                                  )),
+                                  const SizedBox(height: 20),
+                                  TextButton(
+                                    onPressed: () {
+                                      Navigator.pushReplacement(
+                                        context,
+                                        MaterialPageRoute(
+                                            builder: (context) =>
+                                                const MyPorto()),
+                                      );
+                                    },
+                                    child: const Text(
+                                      'Kembali ke Portofolio',
                                       style: TextStyle(
-                                          fontSize: 15, color: Colors.white),
-                                    )),
-                                    const SizedBox(height: 20),
-                                    TextButton(
-                                      onPressed: () {
-                                        Navigator.pushReplacement(
-                                          context,
-                                          MaterialPageRoute(
-                                              builder: (context) =>
-                                                  const MyPorto()),
-                                        );
-                                      },
-                                      child: const Text(
-                                        'Kembali ke Portofolio',
-                                        style: TextStyle(
-                                            color: Color.fromARGB(
-                                                255, 150, 252, 3),
-                                            fontSize: 12),
-                                      ),
-                                    )
-                                  ]),
-                            ),
+                                          color: Color.fromARGB(
+                                              255, 150, 252, 3),
+                                          fontSize: 12),
+                                    ),
+                                  )
+                                ]),
                           );
                         });
                   },
